refactor: rename routes to support new flatlist feature
