+++ app/(root)/_layout.tsx
@@ -18,7 +18,7 @@ export default function Root() {
         }}
       />
       <Stack.Screen
-        name="flatlist"
+        name="listTwoColumns"
         options={{
           headerShown: false,
         }}
